requirements.txt
langchain==0.3.26
langchain-core==0.3.66
langchain-community==0.3.26
langchain-huggingface==0.3.0
huggingface_hub==0.30.2
faiss-cpu==1.8.0
unstructured==0.12.6
pypdf==4.2.0
tiktoken==0.7.0
python-dotenv==1.1.0
